Renamed hash_helper to hash, updated require on pull.js and added sync data helper .. to simplify file_record
