+++ helper/hash.js
